+++ src/lib/graph/VectorLogicalNode.cc
@@ -89,7 +89,7 @@ bool VectorLogicalNode::checkParentValues(unsigned int chain) const
 				     unsigned int chain) const
     {
 	auto par = parents();
-	for (unsigned int i = 0; i < par.size(); ++i) {
+	for (unsigned long i = 0; i < par.size(); ++i) {
 	    if (par[i] == arg) {
 		_func->gradient(grad, _parameters[chain], _lengths, i);
 	    }
